requirements.txt
requests
psycopg2
lxml
cssselect
tqdm
attrs
bs4
python-slugify
google-api-python-client
python-twitter
sentry-sdk==0.5.5